backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1

langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.6
langchain-huggingface==0.0.3

faiss-cpu==1.8.0.post1
pypdf==4.3.1
